+++ voxforge2sphinx/corpus/speaker.py
@@ -26,6 +26,29 @@ import shutil as sh
 import os
 import re
 
+def not_implemented(f):
+    '''A decorator that raises NotImplementedError exception when called.
+    Keeps the docstring of the original function.
+    '''
+    def decorated(*args, **kw):
+        message = 'The method %s() is not implemented yet.' % f.func_name
+        raise NotImplementedError(message)
+    decorated.__doc__ == f.__doc__
+    return decorated
+
+def io_access(f):
+    '''A decorator that handles IOError for functions that access files and directories.
+    Keeps the docstring of the original function.
+    '''
+    def decorated(*args, **kw):
+        try:
+            f(*args, **kw)
+        except IOError as e:
+            print('I/O error(%s): %s %s' % (e.errno, e.strerror, e.filename))
+
+    decorated.__doc__ == f.__doc__
+    return decorated
+
 spkinfo_patterns = [
     ('USERNAME',  r'(?<=User Name:).*(?=\n)'),
     ('GENDER',  r'(?<=Gender:).*(?=\n)'),
@@ -42,46 +65,43 @@ class Speaker:
         self.source = srcdir
         self.target = trgdir
         self.prompts = self.info = self.audios = None
-        self._load()
-
-    def _load(self):
-        '''Load the speaker data.'''
-        try:
-            self._gatherspkinfo()
-            self._gatherprompts()
-            self._gatheraudios()
-        except IOError as e:
-            print('I/O error(%s): %s %s' % (e.errno, e.strerror, e.filename))
 
     def _gatherspkinfo(self):
-        '''Gather all info about that speaker.'''
+        '''Gather all info about that speaker.
+        '''
         filepath = os.path.join(self.source, 'etc', 'README')
         with open(filepath, mode='r', encoding='utf-8') as f:
             tinfo = f.read()
         self.info = {m.lastgroup:m.group(m.lastgroup) for m in spkinfo_p.finditer(tinfo)}
 
+    @io_access
     def _gatherprompts(self):
-        '''Gather all the transcriptions for that speaker.'''
+        '''Gather all the transcriptions for that speaker.
+        '''
         filepath = os.path.join(self.source, 'etc', 'prompts-original')
         with open(filepath, mode='r', encoding='utf-8') as f:
             promptlist =  [(line.split('\n')[0]).split(maxsplit=1) for line in f.readlines()]
-        self.prompts = {prompt[0]:prompt[1] for prompt in promptlist}
+        self.prompts = {prompt[0]:prompt[1] for prompt in promptlist if len(prompt) > 0}
 
+    @not_implemented
     def _gatheraudios(self):
-        '''Gather all the speech audios for that speaker.'''
-        pass
+        '''Gather all the speech audios for that speaker.
+        '''
 
+    @not_implemented
     def export(self, target):
-        '''Save the speaker data in the target folder.'''
-        pass
+        '''Save the speaker data in the target folder.
+        '''
 
     def _copywav(self, srcwavname, trgwavid):
-        '''Copy a wav file to the target directory.'''
+        '''Copy a wav file to the target directory.
+        '''
         srcpath = os.path.join(self.source, 'wav', '%s.wav' % srcwavname)
         trgpath = os.path.join(self.target, 'wav', '%04d' % self.id, '%04d_%03d.wav' % (self.id, trgwavid))
         if not os.path.exists(trgpath):
             sh.copy2(srcpath, trgpath)
 
+    @not_implemented
     def _saveinfo(self):
-        '''Save information to some target file.'''
-        pass
+        '''Save information to some target file.
+        '''
